fix: rm extra file
